Fixed bug with "" empty states
- When states were empty — as in “” or “  “, etc — an error would occur
whereupon no leading time existed for the empty states but the function
still ran and outputted an answer, which resulted in “next day
shipping” or rather a lead-time of 1.
- This happened because originally the comparator function was written
only evaluating for “null” state values because it was wrongly assumed
that if not null then a valid state name would be entered.
- this was resolved by reducing the comparator down to the simplest “
if (value) then” rather than the “if (value is not null or not empty)
then”.  This fixed the bug.